Refactor: Routes 디렉토리명을 pages로 변경, 하이라이트 path 변경
